pyproject.toml
[project]
name = "scholarshare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "anthropic>=0.52.1",
    "gradio>=5.32.0",
    "jinja2>=3.1.6",
    "langchain>=0.3.25",
    "langgraph>=0.4.7",
    "markdown>=3.8",
    "mistralai>=1.8.1",
    "openai>=1.82.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
]
